requirements.txt
python-dotenv
swibots
pymongo
aiohttp
regex